+++ utility/util.h
@@ -10,7 +10,7 @@ struct file_reader {
     file_reader (const char *filename, const char *mode);
     ~file_reader();
     bool read_char(unsigned char &c);
-    template <typename T> bool read_block(int size, std::vector <T> &text);
+    template <typename T> bool read_block(size_t size, std::vector <T> &text);
     template <typename T> void write_block(std::vector<T> &text, const char *sep);
     void reopen();
 private:
